Code fixes, Telemetry issue & Post+update API fix

+++ Application.SmartCharging.BL/adaptor/GroupService.cs
@@ -16,11 +16,11 @@ namespace Application.SmartCharging.BL
 {
     public class GroupService : IGroupService
     {
-
         private readonly IGroupRepository _groupRepository;
         private readonly IConfiguration _configuration;
         private readonly ITelemetryAdaptor _telemetryAdaptor;
         private readonly IMapper _mapper;
+        private readonly CommonValidations _commonValidations;
 
         public GroupService(IConfiguration configuration, ITelemetryAdaptor telemetryAdaptor , IMapper mapper ,IGroupRepository groupRepository)
         {
@@ -28,44 +28,99 @@ namespace Application.SmartCharging.BL
             _configuration = configuration;
             _telemetryAdaptor = telemetryAdaptor;
             _mapper = mapper;
+            _commonValidations = new CommonValidations(groupRepository, telemetryAdaptor);
         }
 
         public async Task<GroupResponse> DeleteGroupAsync(string id)
         {
-            var result = await _groupRepository.DeleteAsync(id);
-            var response = _mapper.Map<GroupResponse>(result);
+            GroupResponse response = new();
+            _telemetryAdaptor.TrackEvent(String.Format("DeleteGroupAsync Started for GroupId {0}", id));
+            try
+            {
+                if (id == null) return null;
+                var result = await _groupRepository.DeleteAsync(id);
+                response = _mapper.Map<GroupResponse>(result);
+            }
+            catch (Exception ex)
+            {
+                _telemetryAdaptor.TrackException(ex);
+            }
+            _telemetryAdaptor.TrackEvent(String.Format("DeleteGroupAsync Completed for GroupId {0}", id));
             return response;
         }
 
         public async Task<IEnumerable<GroupResponse>> GetAllGroupAsync()
         {
-           var result =  await _groupRepository.GetAllAsync();
-           var response = _mapper.Map<IEnumerable<GroupResponse>>(result);
+            IEnumerable<GroupResponse> response = new List<GroupResponse>();
+            _telemetryAdaptor.TrackEvent(String.Format("GetAllGroupAsync Started for GroupId"));
+            try
+            {
+                var result = await _groupRepository.GetAllAsync();
+                 response = _mapper.Map<IEnumerable<GroupResponse>>(result);
+            }
+            catch (Exception ex)
+            {
+                _telemetryAdaptor.TrackException(ex);
+            }
+            _telemetryAdaptor.TrackEvent(String.Format("GetAllGroupAsync Completed for GroupId"));
             return response;
         }
 
         public async Task<GroupResponse> GetGroupAsync(string id)
         {
-            var result = await _groupRepository.GetGroupAsync(id);
-            var response = _mapper.Map<GroupResponse>(result);
+            GroupResponse response = new();
+            _telemetryAdaptor.TrackEvent(String.Format("GetGroupAsync Started for GroupId {0}", id));
+            try
+            {
+                var result = await _groupRepository.GetGroupAsync(id);
+                response = _mapper.Map<GroupResponse>(result);
+            }
+            catch (Exception ex)
+            {
+
+                _telemetryAdaptor.TrackException(ex);
+            }
+            _telemetryAdaptor.TrackEvent(String.Format("GetGroupAsync Completed for GroupId {0}", id));
             return response;
         }
 
         public async Task<GroupResponse> PostGroupAsync(GroupRequest item)
         {
-            var itemToPost = _mapper.Map<Group>(item);
-            var result = await _groupRepository.PostAsync(itemToPost);
-            var response = _mapper.Map<GroupResponse>(result);
+            GroupResponse response = new();
+            _telemetryAdaptor.TrackEvent(String.Format("PostGroupAsync Started"));
+            try
+            {
+                var itemToPost = _mapper.Map<Group>(item);
+                itemToPost.Id= Guid.NewGuid();
+                var result = await _groupRepository.PostAsync(itemToPost);
+                response = _mapper.Map<GroupResponse>(result);
+            }
+            catch (Exception ex)
+            {
+                _telemetryAdaptor.TrackException(ex);
+            }
+            _telemetryAdaptor.TrackEvent(String.Format("PostGroupAsync Completed"));
             return response;
         }
 
-        public async Task<GroupResponse> UpdateGroupAsync(GroupRequest item)
+        public async Task<GroupResponse> UpdateGroupAsync(GroupRequest item, string groupId)
         {
-            
-            var itemToUpdate = _mapper.Map<Group>(item);
-            var result = await _groupRepository.UpdateAsync(itemToUpdate);
-            var response = _mapper.Map<GroupResponse>(result);
+            GroupResponse response = new();
+            _telemetryAdaptor.TrackEvent(String.Format("UpdateGroupAsync Started for groupId {0}" , groupId));
+            try
+            {
+                var itemToUpdate = _mapper.Map<Group>(item);
+                itemToUpdate.Id = Guid.Parse(groupId);
+                var result = await _groupRepository.UpdateAsync(itemToUpdate);
+                 response = _mapper.Map<GroupResponse>(result);
+            }
+            catch (Exception ex)
+            {
+                _telemetryAdaptor.TrackException(ex);
+            }
+            _telemetryAdaptor.TrackEvent(String.Format("UpdateGroupAsync Completed for groupId {0}", groupId));
             return response;
+           
         }
     }
 }
